2.1.0-M1

+++ Server-Core/src/main/java/net/rwhps/server/util/game/GameStartInit.kt
@@ -12,7 +12,8 @@ package net.rwhps.server.util.game
 import net.rwhps.server.data.global.Data
 import net.rwhps.server.net.HttpRequestOkHttp
 import net.rwhps.server.plugin.internal.hess.service.data.HessClassPathProperties
-import net.rwhps.server.util.GameModularLoadClass
+import net.rwhps.server.util.classload.GameModularLoadClass
+import net.rwhps.server.util.classload.GameModularReusableLoadClass
 import net.rwhps.server.util.compression.CompressionDecoderUtils
 import net.rwhps.server.util.file.FileUtil
 import net.rwhps.server.util.log.Log
@@ -24,55 +25,67 @@ import java.lang.reflect.Method
  * @author RW-HPS/Dr
  */
 object GameStartInit {
-    /** Res的MD5 */
-    private const val resMd5 = "6d61d95d9fd7ef679d0013efad1466de"
-    /** Assets的MD5 */
-    private const val assetsMd5 = "b594e7e8d2a0ad925c8ac0e00edbdbad"
+    private enum class ResMD5(val md5: String) {
+        Res("6d61d95d9fd7ef679d0013efad1466de"),
+        Assets("b594e7e8d2a0ad925c8ac0e00edbdbad"),
+        GameModularReusableClass("81b24654be7578fdb56fecf5254a78cb"),
+    }
 
-    fun init(load: GameModularLoadClass): Boolean {
+    fun init(load: GameModularReusableLoadClass): Boolean {
         try {
-            // 加载游戏依赖
-            CompressionDecoderUtils.zipStream(GameStartInit::class.java.getResourceAsStream("/libs.zip")!!).use {
-                it.getSpecifiedSuffixInThePackage("jar",true).each { _, v ->
-                    load.addSourceJar(v)
-                }
-            }
-
-
             val resFile = FileUtil.getFolder(Data.Plugin_GameCore_Data_Path).toFile("Game-Res.zip")
             val assetsFile = FileUtil.getFolder(Data.Plugin_GameCore_Data_Path).toFile("Game-Assets.zip")
+            val gameModularReusableClassFile = FileUtil.getFolder(Data.Plugin_GameCore_Data_Path).toFile("GameModularReusableClass.bin")
             val temp = FileUtil.getFolder(Data.Plugin_GameCore_Data_Path,true)
 
             /* 鉴别两个文件的MD5, 不相同则重下 */
-            if (resFile.md5 != resMd5) {
+            if (resFile.md5 != ResMD5.Res.md5) {
                 resFile.file.delete()
             }
-            if (assetsFile.md5 != assetsMd5) {
+            if (assetsFile.md5 != ResMD5.Assets.md5) {
                 assetsFile.file.delete()
             }
+            if (gameModularReusableClassFile.md5 != ResMD5.GameModularReusableClass.md5) {
+                gameModularReusableClassFile.file.delete()
+            }
 
-            val resTask: (FileUtil,Boolean)->Unit = { file, res ->
-                (if (res) "res" else "assets").let { resName ->
-                    if (!file.exists()) {
+            val resTask: (FileUtil,String,Boolean)->Unit = { file, resName, unzip ->
+                if (!file.exists()) {
+                    if (unzip) {
                         temp.toFolder(resName).file.delete()
+                    }
 
-                        HttpRequestOkHttp.downUrl(Data.urlData.readString("Get.Res")+file.name, file.file, true).also {
-                            Log.clog("$resName : {0}",it)
+                    HttpRequestOkHttp.downUrl(Data.urlData.readString("Get.Res")+file.name, file.file, true).also {
+                        Log.clog("$resName : {0}",it)
+                    }
+                }
+                if (unzip && !temp.toFolder(resName).toFile("Check").exists()) {
+                    CompressionDecoderUtils.zip(file.file).use {
+                        it.getZipAllBytes().each { filePath, bytes ->
+                            temp.toFile(filePath).writeFileByte(bytes)
                         }
                     }
-                    if (!temp.toFolder(resName).toFile("Check").exists()) {
-                        CompressionDecoderUtils.zip(file.file).use {
-                            it.getZipAllBytes().each { filePath, bytes ->
-                                temp.toFile(filePath).writeFileByte(bytes)
-                            }
-                        }
-                        temp.toFolder(resName).toFile("Check").mkdir()
-                    }
+                    temp.toFolder(resName).toFile("Check").mkdir()
                 }
             }
 
-            resTask(resFile,true)
-            resTask(assetsFile,false)
+            resTask(resFile, "res", true)
+            resTask(assetsFile, "assets", true)
+
+            if (GameStartInit::class.java.getResourceAsStream("/libs.zip") == null && gameModularReusableClassFile.notExists()) {
+                resTask(gameModularReusableClassFile, "gameModularReusableClassFile", false)
+                load.readData(gameModularReusableClassFile)
+            } else if (gameModularReusableClassFile.exists()) {
+                load.readData(gameModularReusableClassFile)
+            } else {
+                // 加载游戏依赖
+                CompressionDecoderUtils.zipStream(GameStartInit::class.java.getResourceAsStream("/libs.zip")!!).use {
+                    it.getSpecifiedSuffixInThePackage("jar", true).each { _, v ->
+                        load.addSourceJar(v)
+                    }
+                }
+                load.saveData(FileUtil.getFolder(Data.Plugin_Data_Path).toFile("GameModularReusableClass.bin"))
+            }
         } catch (e: Exception) {
             Log.fatal(e)
             return false
